+++ salvia_rb/lib/salvia_rb/helpers.rb
@@ -1,0 +1,8 @@
+# frozen_string_literal: true
+
+module Salvia
+  module Helpers
+    include Tag
+    include Htmx
+  end
+end